refactor(frontend): relocate models to pages/{domain}/models/
Domain-specific models moved from core/models/ to their respective
pages/{domain}/models/ directories. Inline interfaces extracted from
components into separate model files. Cross-domain models (ApplicationStatus,
CreateApplicationRequest, DimensionResult) remain in core/models/.
