+++ frontend/src/app/pages/ingestion/models/fetch-response.model.ts
@@ -1,0 +1,6 @@
+import { NormalizedJob } from './normalized-job.model';
+
+export interface FetchResponse {
+  count: number;
+  jobs: NormalizedJob[];
+}
