+++ frontend/src/app/pages/ingestion/models/normalized-job.model.ts
