+++ frontend/src/app/pages/ingestion/models/portal-entry.model.ts
