+++ frontend/src/app/pages/ingestion/models/scan-portals-request.model.ts
