+++ frontend/src/app/pages/ingestion/models/scan-result.model.ts
