+++ frontend/src/app/pages/interview/models/competency.model.ts
