+++ frontend/src/app/pages/interview/models/story.model.ts
