+++ frontend/src/app/pages/matching/models/evaluation-result.model.ts
@@ -1,4 +1,4 @@
-import { DimensionResult } from './dimension-result.model';
+import { DimensionResult } from '../../../core/models/dimension-result.model';
 
 export interface EvaluationResult {
   composite_score: number;
