+++ frontend/src/app/pages/matching/models/score-breakdown.model.ts
@@ -1,0 +1,6 @@
+export interface ScoreBreakdown {
+  semantic_score: number;
+  skill_score: number;
+  experience_score: number;
+  language_score: number;
+}
