+++ frontend/src/app/pages/optimization/models/section-change.model.ts
@@ -1,0 +1,6 @@
+export interface SectionChange {
+  section_name: string;
+  original: string;
+  optimized: string;
+  reason: string;
+}
