+++ frontend/src/app/pages/profile/models/cv-section.model.ts
@@ -1,0 +1,4 @@
+export interface CVSection {
+  name: string;
+  content: string;
+}
