+++ frontend/src/app/pages/tracking/models/batch-evaluation-response.model.ts
