+++ frontend/src/app/pages/tracking/models/company-research.model.ts
